Updates:
- Comments updated

Refactor:

- TargetRatio type changed (float -> int) and because of that the calculation too. It's easier now.
- sumVariantRatios method removed the Test model contract.

+++ app/Contracts/ABTest/Model/VariantContract.php
@@ -4,7 +4,17 @@ namespace App\Contracts\ABTest\Model;
 
 interface VariantContract
 {
+    /**
+     * Returns the variant's name
+     *
+     * @return string
+     */
     public function getName(): string;
 
-    public function getTargetRatio(): float;
+    /**
+     * Returns the variant's target ratio
+     *
+     * @return int
+     */
+    public function getTargetRatio(): int;
 }
